beginning of recipe search list build; no api yet
