+++ frontend/src/components/MainSearch/MainSearch.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import './MainSearch.css'
 
+
 class MainSearch extends Component {
 
   render (){
