+++ frontend/src/containers/Homepage/Homepage.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import Header from '../../components/Header/Header.js'
 import Footer from '../../components/Footer/Footer.js';
 import MainSearch from '../../components/MainSearch/MainSearch';
+import RecipeSearchList from '../../components/RecipeSearchList/RecipeSearchList.js';
 
 
 class Homepage extends Component {
@@ -16,7 +17,8 @@ class Homepage extends Component {
       return (
         <div className="wrapper">
           <Header/>
-          <MainSearch/>
+          <MainSearch />
+          <RecipeSearchList />
           <Footer/>
         </div>
       )
